+++ components/utils/data.js
@@ -10,6 +10,7 @@ import {
   // Zstd,
 } from 'numcodecs'
 
+import { sanitizeUrl } from './url'
 import { PROJECTIONS, ASPECTS } from '../constants'
 
 const COMPRESSORS = {
@@ -796,4 +797,106 @@ export const validatePoint = ([lon, lat]) => {
   }
 
   return true
+}
+
+// Infer axes from pyramid consolidated metadata
+const inferCfAxes = (metadata) => {
+  return Object.keys(metadata)
+    .map((k) => {
+      if (!k.startsWith('0/') || !k.endsWith('.zattrs')) {
+        return false
+      }
+
+      const [variable] = k.match(/(?<=0\/)(\w+)(?=\/\.zattrs)/g) ?? []
+
+      if (!variable) {
+        return false
+      }
+
+      const dims = metadata[k]['_ARRAY_DIMENSIONS']
+      if (!dims) {
+        return false
+      }
+
+      const time = dims.find((dim) => metadata[`0/${dim}/.zattrs`]?.calendar)
+      const base = { variable, ...(time ? { T: time } : {}) }
+      if (['x', 'y'].every((d) => dims.includes(d))) {
+        return { ...base, X: 'x', Y: 'y' }
+      } else if (['lat', 'lon'].every((d) => dims.includes(d))) {
+        return { ...base, X: 'lon', Y: 'lat' }
+      }
+    })
+    .filter(Boolean)
+    .reduce((accum, { variable: v, ...rest }) => {
+      accum[v] = rest
+      return accum
+    }, {})
+}
+
+export const inspectDataset = async (url) => {
+  // fetch zmetadata to figure out compression and variables
+  const sanitized = sanitizeUrl(url)
+
+  let response
+  try {
+    response = await fetch(`${sanitized}/.zmetadata`)
+  } catch (e) {
+    // Show generic error message when request fails before response can be inspected.
+    throw new Error(
+      'A network error occurred. This could be a CORS issue or a dropped internet connection.'
+    )
+  }
+
+  if (!response.ok) {
+    const statusText = response.statusText ?? 'Dataset request failed.'
+    if (response.status === 403) {
+      throw new Error(
+        `STATUS 403: Access forbidden. Ensure that URL is correct and that dataset is publicly accessible.`
+      )
+    } else if (response.status === 404) {
+      throw new Error(
+        `STATUS 404: ${statusText} Ensure that URL path is correct.`
+      )
+    } else {
+      throw new Error(
+        `STATUS ${response.status}: ${statusText}. URL: ${sanitized}`
+      )
+    }
+  }
+  const metadata = await response.json()
+
+  if (!metadata.metadata) {
+    throw new Error(metadata?.message || 'Unable to parse metadata')
+  }
+
+  const multiscales = metadata.metadata['.zattrs']['multiscales']
+  let cf_axes = metadata.metadata['.zattrs']['ncviewjs:cf_axes']
+  const rechunking = metadata.metadata['.zattrs']['ncviewjs:rechunking'] ?? []
+
+  if (!multiscales && !cf_axes) {
+    throw new Error('Missing CF axes information')
+  }
+
+  let pyramid = false
+  let visualizedUrl = sanitized
+
+  if (multiscales) {
+    pyramid = true
+
+    cf_axes = inferCfAxes(metadata.metadata)
+
+    if (Object.keys(cf_axes).length === 0) {
+      throw new Error('Cannot infer CF axes information from pyramid')
+    }
+  } else if (rechunking) {
+    const pyramidRechunked = rechunking.find(
+      (r) => r.use_case === 'multiscales'
+    )
+    if (pyramidRechunked) {
+      pyramid = true
+      visualizedUrl = pyramidRechunked.path
+    }
+  }
+
+  return { url: visualizedUrl, cf_axes, metadata, pyramid }
 }
